+++ src/BackerymailsEntityAccessControlHandler.php
@@ -20,10 +20,8 @@ class BackerymailsEntityAccessControlHandler extends EntityAccessControlHandler 
   protected function checkAccess(EntityInterface $entity, $operation, AccountInterface $account) {
     /** @var \Drupal\backerymails\Entity\BackerymailsEntityInterface $entity */
     switch ($operation) {
-      case 'view':
-        return AccessResult::allowedIfHasPermission($account, 'administer backerymails');
-
       case 'delete':
+      case 'view':
         return AccessResult::allowedIfHasPermission($account, 'administer backerymails');
     }